# docs/swagger.yaml
basePath: /api
definitions:
  handlers.AbsensiHistoryItem:
    properties:
      deskripsi:
        type: string
      hari:
        type: string
      id_absen:
        type: integer
      jenis_sholat:
        type: string
      status:
        type: string
      tanggal:
        type: string
    type: object
...
  handlers.AbsensiStaffItem:
    properties:
      deskripsi:
        type: string
      hari:
        type: string
      id_absen:
        type: integer
      jenis_sholat:
        type: string
      jurusan:
        type: string
      kelas:
        type: string
      nama_siswa:
        type: string
      nis:
        type: string
      status:
        type: string
      tanggal:
        type: string
    type: object
...
  handlers.ErrorResponse:
    properties:
      error: {}
      message:
        type: string
    type: object
...
  handlers.HistoryFilters:
    properties:
      end_date:
        type: string
      jurusan:
        type: string
      kelas:
        type: string
      nis:
        type: string
      start_date:
        type: string
      status:
        type: string
    type: object
  handlers.HistorySiswaData:
    properties:
      absensi:
        items:
          $ref: '#/definitions/handlers.AbsensiHistoryItem'
        type: array
      end_date:
        type: string
      periode:
        type: string
      siswa:
        $ref: '#/definitions/handlers.SiswaInfo'
      start_date:
        type: string
      statistik:
        $ref: '#/definitions/handlers.HistoryStatistik'
    type: object
  handlers.HistorySiswaResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.HistorySiswaData'
      message:
        type: string
    type: object
  handlers.HistoryStaffData:
    properties:
      absensi:
        items:
          $ref: '#/definitions/handlers.AbsensiStaffItem'
        type: array
      filters:
        $ref: '#/definitions/handlers.HistoryFilters'
      pagination:
        $ref: '#/definitions/handlers.PaginationInfo'
      statistik:
        $ref: '#/definitions/handlers.LaporanStatistik'
    type: object
  handlers.HistoryStaffResponse:
    properties:
      data:
        $ref: '#/definitions/handlers.HistoryStaffData'
      message:
        type: string
    type: object
  handlers.HistoryStatistik:
    properties:
      persentase_kehadiran:
        type: number
      total_absensi:
        type: integer
      total_alpha:
        type: integer
      total_hadir:
        type: integer
      total_izin:
        type: integer
      total_sakit:
        type: integer
    type: object
...
  handlers.LaporanStatistik:
    properties:
      persentase_alpha:
        type: number
      persentase_hadir:
        type: number
      persentase_izin:
        type: number
      persentase_sakit:
        type: number
      rata_rata_kehadiran:
        type: number
      total_absensi:
        type: integer
      total_alpha:
        type: integer
      total_hadir:
        type: integer
      total_izin:
        type: integer
      total_sakit:
        type: integer
      total_siswa:
        type: integer
    type: object
...
  handlers.PaginationInfo:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total_items:
        type: integer
      total_pages:
        type: integer
    type: object
...
  handlers.SiswaInfo:
    properties:
      jurusan:
        type: string
      kelas:
        type: string
      nama_siswa:
        type: string
      nis:
        type: string
    type: object
...
info:
  contact:
    name: API Support
  description: API untuk sistem pencatatan absensi sholat siswa
  license:
    name: MIT
  termsOfService: http://swagger.io/terms/
  title: API Sistem Absensi Sholat
  version: 1.0.0
paths:
...
  /history/siswa:
    get:
      consumes:
      - application/json
      description: Mengambil riwayat absensi sholat untuk siswa yang sedang login.
        Dapat dilihat per minggu
      parameters:
      - description: 'Minggu ke berapa (0 = minggu ini, 1 = minggu lalu, dst). Default:
          0'
        in: query
        name: week
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Riwayat absensi berhasil diambil
          schema:
            $ref: '#/definitions/handlers.HistorySiswaResponse'
        "401":
          description: Tidak terotentikasi
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "404":
          description: Siswa tidak ditemukan
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Kesalahan server internal
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil riwayat absensi siswa (per minggu)
      tags:
      - history
  /history/staff:
    get:
      consumes:
      - application/json
      description: Mengambil riwayat absensi semua siswa dengan filter. Hanya untuk
        admin, guru, dan wali_kelas
      parameters:
      - description: 'Tanggal mulai (format: YYYY-MM-DD)'
        in: query
        name: start_date
        type: string
      - description: 'Tanggal akhir (format: YYYY-MM-DD)'
        in: query
        name: end_date
        type: string
      - description: Filter berdasarkan kelas
        in: query
        name: kelas
        type: string
      - description: Filter berdasarkan jurusan
        in: query
        name: jurusan
        type: string
      - description: Filter berdasarkan NIS siswa tertentu
        in: query
        name: nis
        type: string
      - description: Filter berdasarkan status (hadir/izin/sakit/alpha)
        in: query
        name: status
        type: string
      - description: 'Nomor halaman (default: 1)'
        in: query
        name: page
        type: integer
      - description: 'Jumlah per halaman (default: 20, max: 100)'
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Riwayat absensi berhasil diambil
          schema:
            $ref: '#/definitions/handlers.HistoryStaffResponse'
        "401":
          description: Tidak terotentikasi
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "403":
          description: Akses ditolak
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
        "500":
          description: Kesalahan server internal
          schema:
            $ref: '#/definitions/handlers.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Ambil riwayat absensi untuk staff (admin/guru/wali_kelas)
      tags:
      - history
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"